+++ public/scripts/movim_visio.js
@@ -34,6 +34,10 @@ var MovimVisio = {
 
         MovimVisio.id = id;
 
+        // Set a lock for the current browser (in case there's several others opened)
+        localStorage.setItem('callId', id);
+        localStorage.setItem('callJid', jid);
+
         let visio = document.querySelector('#visio');
         delete visio.dataset.type;
         visio.dataset.jid = jid;
@@ -293,11 +297,6 @@ var MovimVisio = {
         MovimVisio.observer.observe(body, { childList: true, subtree: true });
     },
 
-    callStart: function (id, jid) {
-        localStorage.setItem('callId', id);
-        localStorage.setItem('callJid', jid);
-    },
-
     callStop: function (id, jid) {
         localStorage.removeItem('callId');
         localStorage.removeItem('callJid');
@@ -309,9 +308,6 @@ Visio_ajaxHttpGetStates();
 MovimWebsocket.attach(() => {
     if (MovimVisio.services.length == 0) {
         Visio_ajaxResolveServices();
-
-        if (localStorage.getItem('callId')) {
-            Visio_ajaxTryForceStop(localStorage.getItem('callId'), localStorage.getItem('callJid'));
-        }
+        Visio_ajaxCheckStatus(localStorage.getItem('callId'), localStorage.getItem('callJid'));
     }
 });