+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v0.33 (master)
 * Implement XEP-0317: Hats and display it in the MUC navigation bar and in a new MUC participant card
 * Jingle and Video-conference fixes
 * Add share_target support in the Manifest and Share widget
+* Handle calls when having multiple browsers connected on one Movim session, allow all the browser to hang-up the current call
 
 v0.32.1
 ---------------------------